Avondwerk. Functionaliteit toegevoegd plus GUI's uitgebreidt.

+++ LPAproject/Artikelen/kano.cs
@@ -9,6 +9,6 @@ namespace LPAproject
     class kano : Boot
     {
         private bool motor = false;
-        public int prijs = 10;
+        public double prijs = 10;
     }
 }
