+++ LPAproject/Artikelen/zeilboot.cs
@@ -9,7 +9,7 @@ namespace LPAproject
     class zeilboot : Boot
     {
         private bool motor = false;
-        public int prijs = 10;
+        public double prijs = 10;
         public string type;
 
         public zeilboot(string Type)
